requirements.txt
python-decouple
swibots
python-telegram
pillow